Creating branch to save work
I had to do some surgery to the student build.gradle file to (1) stop pulling in Barista from our standalone espresso repo, and (2) pull in a newer version of Barista that is compatible with the libraries pulled in by uiautomator.

(Man, it would be a lot more convenient if that standalone espresso lib was pulled back into canvas-android.)

Got a working uiautomator test up and running.  But it doesn't correctly handle the scenario where we launch the student app / student view from the teacher app.  Something about a security exception / untrusted UID.